Rename `coal` to `coalescent`.
